+++ options/tranquility_options.js
@@ -41,7 +41,7 @@ function saveOptions(e) {
         tranquility_font_size:                  document.getElementById("tranquility_font_size").value,
         tranquility_reading_width:              document.getElementById("tranquility_reading_width").value,
         tranquility_line_height:                document.getElementById("tranquility_line_height").value,
-        tranquility_text_align:                 document.getElementById("tranquility_text_align").value    
+        tranquility_text_align:                 document.getElementById("tranquility_text_align").value
     };
 
     // First save all of the current settings in the individual options
@@ -205,7 +205,7 @@ function restoreOptions() {
                         document.getElementById(elem_name).value = result.tranquility_line_height || "140";
                     }
                     else if (opt_name == "tranquility_text_align") {
-                        document.getElementById(elem_name).value = result.tranquility_text_align;
+                        document.getElementById(elem_name).value = result.tranquility_text_align || "left";
                     }
                     else if (opt_name == "tranquility_browser_action_icon") {
                         document.getElementById(elem_name).value = result.tranquility_browser_action_icon;